Перенес эндпоинт фидбэка в отдельный роутер
