+++ src/features_v1/__init__.py
@@ -1,4 +1,5 @@
 from .company_moderator_management.endpoints import router as company_moderator_management_router
+from .company_feedback_router.endpoints import router as company_feedback_router
 from .company_problem_discussion.endpoints import router as company_problem_discussion_router
 from .company_problem_management.endpoints import router as company_problem_management_router
 from .company_problem_meetings.endpoints import router as company_problem_meetings_router
@@ -12,8 +13,10 @@ from .tabit_company_management.endpoints import router as tabit_company_manageme
 from .tabit_department_management.endpoints import router as tabit_department_management_router
 from .tabit_license_management.endpoints import router as tabit_license_management_router
 
+
 __all__ = [
     'company_moderator_management_router',
+    'company_feedback_router',
     'company_problem_discussion_router',
     'company_problem_management_router',
     'company_problem_meetings_router',
